# openapi.json
{"openapi":"3.1.0","info":{"title":"GameVault AI","version":"0.1.0"},"paths":{"/games":{"post":{"summary":"Cadastrar um novo jogo","description":"Permite cadastrar um novo jogo no banco de dados com base nas informações fornecidas.","operationId":"create_game_games_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Listar os jogos","description":"Retorna uma lista de jogos cadastrados no banco de dados com base nos filtros selecionados.","operationId":"list_games_games_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/GameStatus"},{"type":"null"}],"title":"Status"}},{"name":"platform","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Platform"}},{"name":"title","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"}},{"name":"favorite","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Favorite"}},{"name":"sort_by","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/SortField"},{"type":"null"}],"title":"Sort By"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GameResponse"},"title":"Response List Games Games Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/search":{"get":{"summary":"Pesquisar jogos na RAWG","description":"Permite pesquisar jogos na API da RAWG com base no titulo do jogo.","operationId":"search_games_games_search_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RawgGame"},"title":"Response Search Games Games Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/stats":{"get":{"summary":"Obter estatísticas dos seus jogos cadastrados","description":"Retorna estatísticas gerais sobre os jogos cadastrados no banco de dados.","operationId":"get_game_stats_games_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameStats"}}}}},"security":[{"HTTPBearer":[]}]}},"/games/{game_id}":{"get":{"summary":"Obter detalhes de um jogo específico","description":"Retorna os detalhes de um jogo específico com base no ID do Banco de Dados fornecido.","operationId":"get_game_games__game_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"integer","title":"Game Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Atualizar informações proprias de um jogo","description":"Permite atualizar as informações de um jogo cadastrado no banco de dados com base no ID do jogo.","operationId":"update_game_games__game_id__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"integer","title":"Game Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Excluir um jogo","description":"Permite excluir um jogo cadastrado no banco de dados com base no ID do jogo.","operationId":"delete_game_games__game_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"integer","title":"Game Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/import/{rawg_id}":{"post":{"summary":"Importar um jogo da RAWG","description":"Permite importar um jogo da API da RAWG para o banco de dados com base no ID da RAWG fornecido.","operationId":"import_game_games_import__rawg_id__post","parameters":[{"name":"rawg_id","in":"path","required":true,"schema":{"type":"integer","title":"Rawg Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/{game_id}/refresh":{"post":{"summary":"Atualizar informações de um jogo com dados da RAWG","description":"Permite atualizar as informações de um jogo cadastrado no banco de dados com base no ID do jogo com dados da RAWG.","operationId":"refresh_game_games__game_id__refresh_post","parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"integer","title":"Game Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/{game_id}/start":{"patch":{"summary":"Iniciar um jogo, toggle","description":"Toggle para iniciar um jogo cadastrado no banco de dados com base no ID do jogo.","operationId":"start_game_games__game_id__start_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"integer","title":"Game Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/{game_id}/complete":{"patch":{"summary":"Completar um jogo, toggle","description":"Toggle para marcar um jogo como completo no banco de dados com base no ID do jogo.","operationId":"complete_game_games__game_id__complete_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"integer","title":"Game Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/{game_id}/toggle-favorite":{"patch":{"summary":"Marcar/desmarcar um jogo como favorito, toggle","description":"Toggle para marcar um jogo como favorito no banco de dados com base no ID do jogo.","operationId":"toggle_favorite_games__game_id__toggle_favorite_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"integer","title":"Game Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth-test":{"get":{"summary":"Auth Test","operationId":"auth_test_auth_test_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/users":{"post":{"summary":"Criar um novo usuário","description":"Permite criar um novo usuário no banco de dados.","operationId":"create_user_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"summary":"Realizar login","description":"Autentica um usuário e retorna um token JWT.","operationId":"login_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"GameCreate":{"properties":{"title":{"type":"string","title":"Title"},"platform":{"type":"string","title":"Platform"}},"type":"object","required":["title","platform"],"title":"GameCreate"},"GameResponse":{"properties":{"title":{"type":"string","title":"Title"},"platform":{"type":"string","title":"Platform"},"id":{"type":"integer","title":"Id"},"status":{"type":"string","title":"Status"},"personal_rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Personal Rating"},"hours_played":{"type":"number","title":"Hours Played"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"},"favorite":{"type":"boolean","title":"Favorite"},"rawg_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rawg Id"},"cover_image":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cover Image"},"release_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Release Date"},"genres":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Genres"},"metacritic_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Metacritic Score"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"}},"type":"object","required":["title","platform","id","status","hours_played","favorite"],"title":"GameResponse"},"GameStats":{"properties":{"total_games":{"type":"integer","title":"Total Games"},"backlog":{"type":"integer","title":"Backlog"},"playing":{"type":"integer","title":"Playing"},"completed":{"type":"integer","title":"Completed"},"dropped":{"type":"integer","title":"Dropped"},"wishlist":{"type":"integer","title":"Wishlist"},"favorite_games":{"type":"integer","title":"Favorite Games"},"total_hours":{"type":"number","title":"Total Hours"},"average_rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Average Rating"}},"type":"object","required":["total_games","backlog","playing","completed","dropped","wishlist","favorite_games","total_hours","average_rating"],"title":"GameStats"},"GameStatus":{"type":"string","enum":["BACKLOG","PLAYING","COMPLETED","DROPPED","WISHLIST"],"title":"GameStatus"},"GameUpdate":{"properties":{"platform":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Platform"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"personal_rating":{"anyOf":[{"type":"number","maximum":10.0,"minimum":0.0},{"type":"null"}],"title":"Personal Rating"},"hours_played":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Hours Played"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"},"favorite":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Favorite"}},"type":"object","title":"GameUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RawgGame":{"properties":{"rawg_id":{"type":"integer","title":"Rawg Id"},"title":{"type":"string","title":"Title"},"cover_image":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cover Image"},"released":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Released"}},"type":"object","required":["rawg_id","title"],"title":"RawgGame","description":"Representa um jogo resumido utilizado no (/games/search) endpoint da RAWG."},"SortField":{"type":"string","enum":["title","created_at","release_date","platform","status"],"title":"SortField"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","email","password"],"title":"UserCreate"},"UserLogin":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserLogin"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["id","username","email"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"Bearer Auth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"security":[{"Bearer Auth":[]}]}